Revert "change db to postgres"
This reverts commit f1c23abf5359d1da177f63865786e21742577e91.

+++ jobmine/db/migrate/20171125204631_add_latitude_to_profiles.rb
@@ -1,5 +1,5 @@
 class AddLatitudeToProfiles < ActiveRecord::Migration[5.1]
   def change
-    add_column :profiles, :latitude, :decimal
+    add_column :profiles, :latitude, :double
   end
 end
