+++ jobmine/db/migrate/20171125204730_add_longitube_to_profiles.rb
@@ -1,5 +1,5 @@
 class AddLongitubeToProfiles < ActiveRecord::Migration[5.1]
   def change
-    add_column :profiles, :longitude, :decimal
+    add_column :profiles, :longitude, :double
   end
 end
