+++ jobmine/db/migrate/20171125214748_add_latitude_to_jobs.rb
@@ -1,5 +1,5 @@
 class AddLatitudeToJobs < ActiveRecord::Migration[5.1]
   def change
-    add_column :jobs, :latitude, :decimal
+    add_column :jobs, :latitude, :double
   end
 end
